+++ src/reporters/index.ts
@@ -1,4 +1,5 @@
 export * from "./cli.js";
 export * from "./markdown.js";
 export * from "./junit.js";
+export * from "./sarif.js";
 export * from "./utils.js";